created widget in utils
