+++ frontend/lib/l10n/app_de.arb
@@ -64,6 +64,7 @@
   "ourteam": "Unser Team",
   "aboutustext": "Du willst der Welt deine Meinungen, Gedanken und News mitteilen - mit all deinen Gefühlen und Rhetoriken - ganz persönlich? \nViele Kunden wünschen sich eine persönlichere Form der Online-Kommunikation, wie sie heute noch nicht geboten wird. Und hier kommt Speaq zum Einsatz! Wir bieten mit Speaq eine innovative Kommunikationsplattform, die Abwechslung von bekannten Plattformen wie Twitter und Facebook bietet.",
   "frontenddeveloper": "Frontend-Entwickler",
-  "backenddeveloper": "Backend-Entwickler"
+  "backenddeveloper": "Backend-Entwickler",
+  "createdPost": "Speaq wurde erstellt."
 
 }
