+++ frontend/lib/pages/post/new_post_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/utils/all_utils.dart';
 import 'package:frontend/widgets/all_widgets.dart';
+import 'package:frontend/widgets/speaq_post_text_field.dart';
 
 class NewPostPage extends StatefulWidget {
   const NewPostPage({Key? key}) : super(key: key);
@@ -11,6 +12,7 @@ class NewPostPage extends StatefulWidget {
 
 class _NewPostPageState extends State<NewPostPage> {
   late Size deviceSize;
+  late AppLocalizations appLocale;
   final TextEditingController _postController = TextEditingController();
   bool emojiShowing = false;
   bool visibilityContainer = false;
@@ -18,33 +20,34 @@ class _NewPostPageState extends State<NewPostPage> {
 
   @override
   void initState() {
-    setState(() {
-      visibilityContainer = checkValue;
-    });
+    setState(
+      () {
+        visibilityContainer = checkValue;
+      },
+    );
   }
 
   @override
   Widget build(BuildContext context) {
     deviceSize = MediaQuery.of(context).size;
-    AppLocalizations appLocale = AppLocalizations.of(context)!;
+    appLocale = AppLocalizations.of(context)!;
     return SafeArea(
       child: Scaffold(
-        appBar: SpqAppBar(
-          preferredSize: deviceSize,
-          actionList: [_buildSendPostButton()],
-        ),
-        body: Column(
-          children: [
-            Expanded(
-              child: _buildPostTextField(appLocale),
-            ),
-            buildVisibilityContainer(),
-            buildMainContainer(deviceSize),
-          ],
-        )
-
+          appBar: SpqAppBar(
+            preferredSize: deviceSize,
+            actionList: [_buildSendPostButton(appLocale)],
+          ),
+          body: Column(
+            children: [
+              Expanded(
+                child: _buildPostTextField(appLocale),
+              ),
+              buildVisibilityContainer(),
+              buildMainContainer(deviceSize),
+            ],
+          )
           //_buildPostTextField(appLocale),
-      ),
+          ),
     );
   }
 
@@ -79,29 +82,30 @@ class _NewPostPageState extends State<NewPostPage> {
 
   Padding buildContainerPictures(String url) {
     return Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: InkWell(
-          onTap: () {
-            print('Hello World');
-          },
-          child: Container(
-            decoration: BoxDecoration(
-                color: Colors.black12, borderRadius: BorderRadius.circular(5)),
-            child: ClipRRect(
-              borderRadius: BorderRadius.circular(5),
-              child: SizedBox.fromSize(
-                size: Size.fromRadius(32), // Image radius
-                child: Image.network(url, fit: BoxFit.cover),
-              ),
+      padding: const EdgeInsets.all(8.0),
+      child: InkWell(
+        onTap: () {
+          print('Hello World');
+        },
+        child: Container(
+          decoration: BoxDecoration(
+              color: Colors.black12, borderRadius: BorderRadius.circular(5)),
+          child: ClipRRect(
+            borderRadius: BorderRadius.circular(5),
+            child: SizedBox.fromSize(
+              size: Size.fromRadius(32), // Image radius
+              child: Image.network(url, fit: BoxFit.cover),
             ),
           ),
-        ));
+        ),
+      ),
+    );
   }
 
-  Material buildMaterial(IconData icon) {
+  Material buildMaterial(IconData icon, void Function()? onPressed) {
     return Material(
       child: IconButton(
-        onPressed: () {},
+        onPressed: onPressed,
         icon: Icon(
           icon,
           color: Colors.black54,
@@ -115,8 +119,8 @@ class _NewPostPageState extends State<NewPostPage> {
       height: deviceSize.height * 0.08,
       child: Row(
         children: [
-          buildMaterial(Icons.emoji_emotions),
-          buildMaterial(Icons.gif_box),
+          buildMaterial(Icons.emoji_emotions, null),
+          buildMaterial(Icons.gif_box, null),
           Expanded(
             child: Padding(
               padding: const EdgeInsets.symmetric(vertical: 8.0),
@@ -158,7 +162,6 @@ class _NewPostPageState extends State<NewPostPage> {
               ),
             ),
           ),
-          buildMaterial(Icons.send),
         ],
       ),
     );
@@ -174,117 +177,32 @@ class _NewPostPageState extends State<NewPostPage> {
         ),
       );
 
-  Widget _buildSendPostButton() => TextButton(
-        onPressed: () => print("Speaq"),
+  Widget _buildSendPostButton(AppLocalizations appLocale) => TextButton(
+        onPressed: () => showDialog(
+          context: context,
+          builder: (BuildContext context) {
+            Future.delayed(
+              Duration(seconds: 1),
+              () {
+                Navigator.popAndPushNamed(context, "home");
+              },
+            );
+            return AlertDialog(
+              title: Text(appLocale.createdPost),
+              backgroundColor: spqLightGrey,
+            );
+          },
+        ),
         child: Container(
           child: const Text("speaq"),
           margin: const EdgeInsets.symmetric(horizontal: 4.0),
           padding: const EdgeInsets.symmetric(vertical: 2.0, horizontal: 4.0),
           decoration: BoxDecoration(
-              border: Border.all(color: spqPrimaryBlue, width: 1.0),
-              borderRadius: const BorderRadius.all(Radius.circular(16.0))),
+            border: Border.all(color: spqPrimaryBlue, width: 1.0),
+            borderRadius: const BorderRadius.all(
+              Radius.circular(16.0),
+            ),
+          ),
         ),
       );
-}
-
-class SpqPostTextField extends StatelessWidget {
-  const SpqPostTextField({
-    Key? key,
-    required this.controller,
-    this.obscureText = false,
-    this.isPassword = false,
-    required this.hintText,
-    this.prefixIcon,
-    this.suffixIcon,
-    this.onTap,
-    this.validator,
-    this.keyboardType,
-    this.minLines,
-    this.maxLines,
-    this.width,
-    this.height = 56,
-    this.contentPadding =
-        const EdgeInsets.symmetric(vertical: 16.0, horizontal: 8.0),
-    this.enabled = true,
-  }) : super(key: key);
-
-  final TextEditingController? controller;
-  final bool obscureText;
-  final bool isPassword;
-  final String hintText;
-  final Widget? prefixIcon;
-  final Widget? suffixIcon;
-  final void Function()? onTap;
-  final TextInputType? keyboardType;
-  final String? Function(String?)? validator;
-  final int? minLines;
-  final int? maxLines;
-  final double? width;
-  final double? height;
-  final EdgeInsets? contentPadding;
-  final bool enabled;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: width,
-      height: height,
-      child: TextFormField(
-        textAlign: TextAlign.start,
-        textAlignVertical: TextAlignVertical.center,
-        minLines: minLines,
-        maxLines: maxLines,
-        obscureText: obscureText,
-        readOnly: isPassword,
-        keyboardType: keyboardType,
-        enableSuggestions: !isPassword,
-        autocorrect: !isPassword,
-        controller: controller,
-        validator: (value) => validator!(value),
-        style: const TextStyle(color: spqBlack, fontSize: 16),
-        enabled: enabled,
-        decoration: InputDecoration(
-          isDense: true,
-          label: Container(
-            margin: const EdgeInsets.only(bottom: 12.0),
-            child: Text(hintText,
-                style: const TextStyle(
-                    color: spqLightGrey, fontWeight: FontWeight.w100)),
-          ),
-          contentPadding: contentPadding,
-          labelStyle:
-              const TextStyle(color: spqLightGrey, fontWeight: FontWeight.w100),
-          floatingLabelBehavior: FloatingLabelBehavior.never,
-          floatingLabelAlignment: FloatingLabelAlignment.start,
-          alignLabelWithHint: true,
-          prefixIcon: prefixIcon != null
-              ? Padding(
-                  padding: const EdgeInsets.all(12.0),
-                  child: prefixIcon,
-                )
-              : null,
-          suffixIcon: suffixIcon,
-          fillColor: spqWhite,
-          filled: true,
-          focusedBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(8.0),
-            borderSide: const BorderSide(color: spqPrimaryBlue, width: 1.0),
-          ),
-          hintStyle: const TextStyle(
-              color: spqLightGrey, fontSize: 16, fontWeight: FontWeight.w100),
-          //hintText: hintText,
-          enabledBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(8.0),
-            borderSide: const BorderSide(color: spqWhite, width: 1.0),
-          ),
-          disabledBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(8.0),
-            borderSide: const BorderSide(color: spqWhite, width: 1.0),
-          ),
-          border: const OutlineInputBorder(),
-        ),
-        onTap: onTap,
-      ),
-    );
-  }
 }
